# docs/rental-api-openapi.yaml
openapi: 3.0.3
info:
  title: Rental Management API — RentFlow
  description: |
    Multi-tenant SaaS API for rental property management in Uganda.

    ## Authentication
    All endpoints except `/auth/register`, `/auth/login`, and `/auth/refresh`
    require a Bearer JWT token:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Token Expiry
    - Access tokens expire after 24 hours.
    - On 401, the frontend Axios interceptor checks if the refresh token is expired.
    - If refresh token is still valid → silent refresh via `/auth/refresh`, retry original request.
    - If refresh token is also expired → `logout()` + redirect to `/login`.

    ## Pagination
    All list endpoints support: `page` (default 0), `size` (default 10),
    `sortBy` (default createdAt), `sortDir` (asc/desc, default desc).

    ## Billing Cycles
    Each agreement has a `billingDay` (1–28) derived automatically from
    `startDate` (day of month, capped at 28 to avoid February issues).
    Payments are recorded against exact cycle date ranges
    (`periodStartDate` to `periodEndDate`) rather than month/year numbers.

    Two billing models per agreement:
    - `ADVANCE` — tenant pays at the start of each cycle
    - `ARREARS` — tenant pays at the end of each cycle

    ## Balance Computation
    Outstanding balance is cumulative across ALL billing cycles, not just the current month.
    Unpaid previous months carry forward automatically.

...

    ## Overpayment
...

    ## Settings
...

...
servers:
...
paths:
...
  /auth/register:
    post:
      tags: [ Auth ]
      summary: Register a new landlord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              name: Brian Nahabwe
              phoneNumber: "256774010511"
              password: secret123
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [ Auth ]
      summary: Login with phone number or email
      description: |
        Returns accessToken, refreshToken, name, phoneNumber, email.
        Frontend stores in authStore (Zustand + localStorage).
        Frontend also immediately fetches GET /settings after login to populate settingsStore.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              phone:
                value: { username: "256774010511", password: secret123 }
              email:
                value: { username: brian@example.com, password: secret123 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/refresh:
    post:
      tags: [ Auth ]
      summary: Refresh access token
      description: |
        Exchange a valid refresh token for a new access token.
        Called automatically by the Axios response interceptor on 401.
        The interceptor checks isRefreshTokenExpired() before calling this —
        if the refresh token is already expired it skips directly to logout().
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ─── TENANTS ───────────────────────────────────────────
...
  /units:
    get:
      tags: [ Units ]
      summary: List all units (paginated + search + availability filter)
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: search
          in: query
          schema:
            type: string
        - name: isAvailable
          in: query
          schema:
            type: boolean
          description: true = available only, false = occupied only, omit = all
      responses:
        '200':
          description: Paginated units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUnitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [ Units ]
      summary: Create a unit
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalUnitRequest'
            example:
              roomNumber: A1
              description: Self-contained, ground floor
              rentAmount: 350000
              isAvailable: true
      responses:
        '201':
          description: Unit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalUnitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /units/{id}:
    get:
      tags: [ Units ]
      summary: Get a unit
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Unit found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalUnitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [ Units ]
      summary: Update a unit
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalUnitRequest'
      responses:
        '200':
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalUnitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [ Units ]
      summary: Delete a unit
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── AGREEMENTS ────────────────────────────────────────
...
  /payments:
    get:
      tags: [ Payments ]
      summary: List all payments (paginated + search + date filter)
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: search
          in: query
          schema:
            type: string
          description: Partial match on tenant name, room number, or reference
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Filter by paymentDate >= from (inclusive). Use with `to`.
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: Filter by paymentDate <= to (inclusive). Use with `from`.
        - name: tenantId
          in: query
          schema:
            type: string
            format: uuid
        - name: agreementId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPaymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [ Payments ]
      summary: Record a payment
      description: |
        Records a payment for a specific billing cycle.

        **Period dates:** Must match the agreement's billing cycle exactly.
        The frontend cycle picker generates these based on `billingDay`.

        **Overpayment:** If `amount > expectedAmount`, excess auto-creates a ROLLOVER
        payment for the next cycle. Chains recursively across months.

        **Split payments:** Multiple payments per cycle allowed. System sums them.

        **Period status** (computed, never stored):
        - PAID — totalPaidForPeriod >= expectedAmount
        - PARTIAL — 0 < totalPaidForPeriod < expectedAmount
        - ROLLOVER — source = ROLLOVER (auto-created by system)
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            examples:
              full_payment_mid_month:
                summary: Full payment — billing_day = 15
                value:
                  agreementId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  paymentDate: "2026-04-15"
                  amount: 180000
                  method: CASH
                  periodStartDate: "2026-04-15"
                  periodEndDate: "2026-05-14"
                  reference: "40422387417"
              full_payment_calendar:
                summary: Full payment — billing_day = 1
                value:
                  agreementId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  paymentDate: "2026-04-01"
                  amount: 180000
                  method: CASH
                  periodStartDate: "2026-04-01"
                  periodEndDate: "2026-04-30"
              overpayment:
                summary: Overpayment — covers two cycles
                value:
                  agreementId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  paymentDate: "2026-04-01"
                  amount: 360000
                  method: CASH
                  periodStartDate: "2026-04-01"
                  periodEndDate: "2026-04-30"
              partial:
                summary: Partial payment
                value:
                  agreementId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  paymentDate: "2026-04-01"
                  amount: 100000
                  method: CASH
                  periodStartDate: "2026-04-01"
                  periodEndDate: "2026-04-30"
                  notes: Balance to follow end of month
      responses:
        '201':
          description: Payment recorded. ROLLOVER auto-created if overpaid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /reports/summary:
    get:
      tags: [ Reports ]
      summary: Dashboard summary statistics
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/payments:
    get:
      tags: [ Reports ]
      summary: Payment report for a date range
      security:
        - BearerAuth: [ ]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
          example: "2026-04-01"
        - name: to
          in: query
          schema:
            type: string
            format: date
          example: "2026-04-30"
      responses:
        '200':
          description: Payment report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/occupancy:
    get:
      tags: [ Reports ]
      summary: Occupancy rate
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Occupancy report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccupancyReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── SETTINGS ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
    SortBy:
      name: sortBy
      in: query
      schema:
        type: string
        default: createdAt
    SortDir:
      name: sortDir
      in: query
      schema:
        type: string
        enum: [ asc, desc ]
        default: desc
  responses:
    BadRequest:
      description: Validation or business rule error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: BAD_REQUEST
            message: Unit A1 already has an active agreement
    Unauthorized:
      description: Missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: UNAUTHORIZED
            message: Invalid or expired token
  schemas:
...
    RegisterRequest:
      type: object
      required: [ name, phoneNumber, password ]
      properties:
        name:
          type: string
          example: Brian Nahabwe
        phoneNumber:
          type: string
          example: "256774010511"
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
          minLength: 6
    LoginRequest:
      type: object
      required: [ username, password ]
      properties:
        username:
          type: string
          description: Phone number or email address
          example: "256774010511"
        password:
          type: string
    RefreshRequest:
      type: object
      required: [ refreshToken ]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token. Expires in 24 hours.
        refreshToken:
          type: string
          description: JWT refresh token. Expires in 30 days.
        name:
          type: string
          example: Brian Nahabwe
        phoneNumber:
          type: string
          example: "256774010511"
        email:
          type: string
          nullable: true

    # ── Tenants ─────────────────────────────────────────
...
    RentalUnitRequest:
      type: object
      required: [ roomNumber, rentAmount ]
      properties:
        roomNumber:
          type: string
          example: A1
        description:
          type: string
          nullable: true
        rentAmount:
          type: number
          format: decimal
          example: 350000
        isAvailable:
          type: boolean
          default: true
    RentalUnitResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        roomNumber:
          type: string
        description:
          type: string
          nullable: true
        rentAmount:
          type: number
          format: decimal
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PagedUnitResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RentalUnitResponse'

    # ── Agreements ──────────────────────────────────────
...
    PaymentMethod:
      type: string
      enum: [ CASH ]
      description: MVP supports CASH only. Mobile money planned for post-MVP.
    PaymentSource:
      type: string
      enum: [ CASH, ROLLOVER ]
      description: |
        CASH = manually recorded by landlord.
        ROLLOVER = auto-created by system from previous cycle overpayment.
    PeriodStatus:
      type: string
      enum: [ PAID, PARTIAL, ROLLOVER ]
      description: Computed dynamically from payment sums. Never stored.
    PaymentRequest:
      type: object
      required: [ agreementId, paymentDate, amount, method, periodStartDate, periodEndDate ]
      properties:
        agreementId:
          type: string
          format: uuid
          description: Must be an ACTIVE agreement. tenantId and unitId derived automatically.
        paymentDate:
          type: string
          format: date
          description: Date payment was physically received.
        amount:
          type: number
          format: decimal
        method:
          $ref: '#/components/schemas/PaymentMethod'
        periodStartDate:
          type: string
          format: date
          description: Start of billing cycle this payment covers. Generated by frontend cycle picker.
          example: "2026-04-15"
        periodEndDate:
          type: string
          format: date
          description: End of billing cycle. One day before next cycle start.
          example: "2026-05-14"
        reference:
          type: string
          nullable: true
          description: Mobile money transaction ID or other reference.
        notes:
          type: string
          nullable: true
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agreementId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        tenantName:
          type: string
        unitId:
          type: string
          format: uuid
        roomNumber:
          type: string
        paymentDate:
          type: string
          format: date
        amount:
          type: number
          format: decimal
        method:
          $ref: '#/components/schemas/PaymentMethod'
        periodStartDate:
          type: string
          format: date
          example: "2026-04-15"
        periodEndDate:
          type: string
          format: date
          example: "2026-05-14"
        expectedAmount:
          type: number
          format: decimal
          description: rentAmount from agreement at time of payment.
        overpayment:
          type: number
          format: decimal
          description: Excess above expectedAmount. 0 if not overpaid.
        source:
          $ref: '#/components/schemas/PaymentSource'
        periodStatus:
          $ref: '#/components/schemas/PeriodStatus'
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PagedPaymentResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/PaymentResponse'

    # ── Reports ─────────────────────────────────────────
    SummaryResponse:
      type: object
      properties:
        totalUnits:
          type: integer
        occupiedUnits:
          type: integer
        availableUnits:
          type: integer
        totalTenants:
          type: integer
        activeAgreements:
          type: integer
        terminatedAgreements:
          type: integer
        totalRevenueAllTime:
          type: number
          format: decimal
          description: Sum of all payment amounts ever recorded.
    PaymentReportResponse:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        totalPayments:
          type: integer
        totalAmount:
          type: number
          format: decimal
    OccupancyReportResponse:
      type: object
      properties:
        totalUnits:
          type: integer
        occupiedUnits:
          type: integer
        availableUnits:
          type: integer
        occupancyRate:
          type: number
          format: decimal
          description: Percentage 0.00–100.00

    # ── Settings ────────────────────────────────────────
...
    PagedResponse:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Unit A1 already has an active agreement