# src/main/java/com/cognix/rentalcoreapi/modules/auth/dto/RefreshRequest.java
package com.cognix.rentalcoreapi.modules.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record RefreshRequest(
        @NotBlank(message = "Refresh token is required")
        String refreshToken
) {
}